adapt to develop changes
